# Make copy, similar and getindex fallback to Array