Fixes #34

It seems that this improves `getindex` timing but reduces `copy` time (which needs to convert the sentinel):

Before this PR:
```
julia> x = SentinelArray(rand(10^6));

julia> @btime copy($x);
  442.980 μs (3 allocations: 7.63 MiB)

julia> @btime similar($x);
  251.363 μs (6 allocations: 7.63 MiB)

julia> @btime $x[10^5:-1:1];
  173.375 μs (6 allocations: 781.50 KiB)
```

This PR:
```
julia> x = SentinelArray(rand(10^6));

julia> @btime copy($x);
  1.412 ms (2 allocations: 8.58 MiB)

julia> @btime similar($x);
  265.543 μs (2 allocations: 8.58 MiB)

julia> @btime $x[10^5:-1:1];
  155.686 μs (2 allocations: 879.02 KiB)
```

Still - I believe that all these operations are one-time only, so even the regression for `copy` is acceptable I think (and `copy` is not that bad - it is quite fast anyway). The reality (as I assume), is that `SentinelArray` will be mostly created only when you read-in the data.

(I would expect that for `ChainedVector` backed `SentinelVector` to see a similar situation)