@quinnj 2020-08-30
Yeah, I'm in favor of this. I spent a decent chunk of time trying to find a way to optimize today, but couldn't manage to come up with anything meaningfully faster.